+++ test/enum_test.exs
@@ -3,6 +3,18 @@ defmodule Moar.EnumTest do
 
   use Moar.SimpleCase, async: true
 
+  doctest Moar.Enum
+
+  describe "at!" do
+    test "like `Moar.Enum.at/3` but raises if the index is out of bounds" do
+      ["A", "B", "C"] |> Moar.Enum.at!(1) |> assert_eq("B")
+
+      assert_raise RuntimeError, ~s|Out of range: index 4 of enum with length 3: ["A", "B", "C"]|, fn ->
+        ["A", "B", "C"] |> Moar.Enum.at!(4)
+      end
+    end
+  end
+
   describe "first!" do
     test "returns the first item in the enum" do
       assert Moar.Enum.first!(["foo"]) == "foo"
@@ -42,6 +54,24 @@ defmodule Moar.EnumTest do
     test "sorts by the mapper case-insensitively" do
       Moar.Enum.isort_by([%{name: "Banana"}, %{name: "apple"}, %{name: "cherry"}], & &1.name)
       |> assert_eq([%{name: "apple"}, %{name: "Banana"}, %{name: "cherry"}])
+    end
+  end
+
+  describe "take_at" do
+    test "returns a list of elements at the given indices" do
+      ["A", "B", "C"] |> Moar.Enum.take_at([1]) |> assert_eq(["B"])
+      ["A", "B", "C"] |> Moar.Enum.take_at([1, 2]) |> assert_eq(["B", "C"])
+      ["A", "B", "C"] |> Moar.Enum.take_at([1, 4]) |> assert_eq(["B", nil])
+    end
+
+    test "when given `:all`, returns the entire list" do
+      ["A", "B", "C"] |> Moar.Enum.take_at(:all) |> assert_eq(["A", "B", "C"])
+    end
+  end
+
+  describe "tids" do
+    test "returns :tid fields" do
+      [%{tid: "a"}, %{tid: nil}, %{tid: "c"}] |> Moar.Enum.tids() |> assert_eq(["a", nil, "c"])
     end
   end
 end